01-DevOps-Projects/Django-GitOps-AWS/app/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.7.0